requirements.txt
git+https://c9b60542c1a793a1bcc0b1e95ce3ab3d2da148c4:x-oauth-basic@github.com/getcircle/protobuf-to-dict.git@0.2.1
git+https://c9b60542c1a793a1bcc0b1e95ce3ab3d2da148c4:x-oauth-basic@github.com/getcircle/protobuf-soa.git@0.5.1
git+https://c9b60542c1a793a1bcc0b1e95ce3ab3d2da148c4:x-oauth-basic@github.com/getcircle/protobuf-registry.git@2.47.1
git+https://c9b60542c1a793a1bcc0b1e95ce3ab3d2da148c4:x-oauth-basic@github.com/getcircle/python-soa.git@0.17.8
protobuf==3.0.0a3
contextlib2==0.4.0
factory-boy==2.4.1
Flask==0.10.1
gevent==1.0.2
greenlet==0.4.9
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.26
locustio==0.7.3
MarkupSafe==0.23
msgpack-python==0.4.6
pytz==2015.6
requests==2.7.0
unpyclib==0.8.1
Werkzeug==0.10.4
# install the load-test lib
-e .
